+++ src/java/com/rapleaf/jack/ModelWithId.java
@@ -17,13 +17,13 @@ package com.rapleaf.jack;
 import java.io.Serializable;
 
 public abstract class ModelWithId implements Serializable {
-  private final int id;
+  private final long id;
 
-  protected ModelWithId(int id) {
+  protected ModelWithId(long id) {
     this.id = id;
   }
 
-  public int getId() {
+  public long getId() {
     return id;
   }
